+++ opentelemetry-exporter-gcp-trace/tests/test_cloud_trace_exporter.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import re
 import unittest
 from unittest import mock
 
@@ -620,6 +621,54 @@ class TestCloudTraceSpanExporter(unittest.TestCase):
 
     def test_extract_empty_resources(self):
         self.assertEqual(_extract_resources(Resource.get_empty()), {})
+
+    def test_extract_resource_attributes_with_regex(self):
+        resource_regex = re.compile(r"service\..*")
+        resource = Resource(
+            attributes={
+                "cloud.account.id": 123,
+                "host.id": "host",
+                "cloud.zone": "US",
+                "cloud.provider": "gcp",
+                "extra_info": "extra",
+                "gcp.resource_type": "gce_instance",
+                "not_gcp_resource": "value",
+                "service.name": "my-app",
+                "service.version": "1",
+            }
+        )
+        expected_extract = {
+            "g.co/r/gce_instance/project_id": "123",
+            "g.co/r/gce_instance/instance_id": "host",
+            "g.co/r/gce_instance/zone": "US",
+            "service.name": "my-app",
+            "service.version": "1",
+        }
+        self.assertEqual(
+            _extract_resources(resource, resource_regex), expected_extract
+        )
+
+    def test_non_matching_regex(self):
+        resource_regex = re.compile(r"this-regex-matches-nothing")
+        resource = Resource(
+            attributes={
+                "cloud.account.id": 123,
+                "host.id": "host",
+                "cloud.zone": "US",
+                "cloud.provider": "gcp",
+                "extra_info": "extra",
+                "gcp.resource_type": "gce_instance",
+                "not_gcp_resource": "value",
+            }
+        )
+        expected_extract = {
+            "g.co/r/gce_instance/project_id": "123",
+            "g.co/r/gce_instance/instance_id": "host",
+            "g.co/r/gce_instance/zone": "US",
+        }
+        self.assertEqual(
+            _extract_resources(resource, resource_regex), expected_extract
+        )
 
     def test_extract_well_formed_resources(self):
         resource = Resource(